move the durable function related extraction logic to durable.py
